QtContacts: be careful about which queries to use
"query-builder" controls how the qtcontacts-tracker backend
does certain things. I can be used to enable a new implementation
which is not the default yet.

For save, the new method is already the default in recent
qtcontacts-tracker, but it doesn't hurt to enable it explicitly.  For
fetch, the new method is not the default, but it is needed by us
because without it, the last modified time stamp is not fetched.

We must *not* use query-builder=all, because the new method of
deleting contacts is known to be broken and doesn't do anything.