config fix: operations on non-peer configs failed (MBC #3157)
When running operations on a non-peer configuration (like --restore @default
addressbook), the operation fails with
[ERROR] <source name>: type 'select backend' not supported

The root cause is that the "type" property was read from a virtual, empty peer
node. MultiplexConfigNode already handled this by falling back to the non-peer
node.

The same special case also needs to be added to SyncSourceNodes::getNode().