KCalExtended: added initial SyncSource
Change tracking in this new source is based on the KCalExtended change
tracking functions. This deviates from the "normal" change tracking in
SyncEvolution (TrackingSyncSource and id/revision pairs) and thus
requires different inhertitance and initialization steps.

Like SyncSourceRevisions, the source needs to delay progress to ensure
that time increments between syncs. In contrast to
SyncSourceRevisions, this is done in a very simplistic way without
considering whether any changes were made. Better improve that
(refactor SyncSourceRevisions?), because it slows down testing. It is
also unclear whether the delay has to happen before or after taking
the time stamp for the next sync.

The code works well enough to pass some Client::Source
tests. Compiling it is a bit tricky because it is unclear how exactly
the KCalExtended header files will be installed. The assumption is
that eventually it'll be in a "kcal" directory.