KCalExtended: must preserve creation time during update
It is the responsibility of the application to ensure that the
right creation time is set for the updated item: preserve the
original one or use the one from the update date, if that is
set.

The correct behavior depends a bit on whether created() maps
to DTSTAMP or CREATED: CREATED has to be preserved unconditionally
(current code in KCalExtended) whereas DTSTAMP is conditional.