KCalExtended: avoid adding incidence anew
addIncidence() accepts an incidence which has the same
UID/RECURRENCE-ID pair as an existing incidence. But later on saving
that modified calendar fails.

It is the responsibility of the KCalExtended user to avoid this
problem, which is what this patch does. This is similar to
EvolutionCalendarSource, except that there adding an item fails more
gracefully with an "exists already" error.