support local sync (BMC #712)
Local sync is configured with a new syncURL = local://<context> where
<context> identifies the set of databases to synchronize with. The
URI of each source in the config identifies the source in that context
to synchronize with.

The databases in that context run a SyncML session as client. The
config itself is for a server. Reversing these roles is possible by
putting the config into the other context.

A sync is started by the server side, via the new LocalTransportAgent.
That agent forks, sets up the client side, then passes messages
back and forth via stream sockets. Stream sockets are useful because
unexpected peer shutdown can be detected.

Running the server side requires a few changes:
- do not send a SAN message, the client will start the
  message exchange based on the config
- wait for that message before doing anything

The client side is more difficult:
- Per-peer config nodes do not exist in the target context.
  They are stored in a hidden .<context> directory inside
  the server config tree. This depends on the new "registering nodes
  in the tree" feature. All nodes are hidden, because users
  are not meant to edit any of them. Their name is intentionally
  chosen like traditional nodes so that removing the config
  also removes the new files.
- All relevant per-peer properties must be copied from the server
  config (log level, printing changes, ...); they cannot be set
  differently.

Because two separate SyncML sessions are used, we end up with
two normal session directories and log files.

The implementation is not complete yet:
- no glib support, so cannot be used in syncevo-dbus-server
- no support for CTRL-C and abort
- no interactive password entry for target sources
- unexpected slow syncs are detected on the client side, but
  not reported properly on the server side