SyncContext: avoid crash in sort() with LogDir as binary predicate
LogDir is not copyable when it contains an m_info pointer. std::sort()
inside LogDir::getLogDirs() for a context is called with LogDir as
binary predicate, which causes a crash because sort() copies the
predicate.

This was found when calling getLogDirs() in a LogDir used for local
sync, without a peer. It did not happen before because then the
LogDir instance did not contain an m_info pointer. It was merely
inefficient.

This patch moves the static log dir name methods into a new
LogDirNames class and uses that as binary predicate. LogDir itself
inherits these methods via a private base class, which is similar
to the previous private static methods (avoids touching all of
the calls) and also inherits from boost::copyable to catch future
errors related to copying LogDir.