runtests.py: make repos group-writable
The usage model for the nightly repos is that the "nightly" user
updates them and other users modify them to try out changes before
committing them. This relies on group-writable repos.

Despite umask 0002 and drwxrwsr-x permissions of the work area
this didn't work, group only had read permission. This patch adds
an explicit chmod and sets "sharedRepository = group" to fix this
after a fresh clone.