Nokia phones: alarm times in UTC, sending PHOTO (BMC #1657, #5860)
Introduced a new remote rule for all devices made by Nokia. In that
rule, we set certain defaults which turned out to be useful for
a variety of existing phones:
- send alarm times in UTC (BMC #1657)
- ignore the incorrect 256 bytes size limit for PHOTOs
  and other properties (BMC #5860)

Should a specific model require different values, then a rule for
it after the initial 00_nokia.xml can override the defaults.