KCalExtended: select notebook (BMC #6898)
"evolutionsource" either refers to a notebook name in default storage, or a
database file path.
1) If it starts with file:// prefix, the dbfile will be created if it does
not exist yet. Default notebook in the database file will be selected.
2) If it starts with SyncEvolution_Test_ prefix, it stands for notebook's
name in default storage, the notebook will be created if it does not exist yet.
3) If no special prefix, search notebook in default storage, if found,
select that notebook, else throw an error.
4) If no evolutionsource configured, default notebook named "Default" will
be used.