syncevo-dbus-server: introduced timeout utility class
The new Timeout class wraps g_timeout_add_seconds() and
g_source_remove() in a C++ interface. Users of the class no longer
have to worry about callbacks with stale data pointers, because the
Timeout class will automatically remove the timeout when it gets
deleted. Additional parameters for the callback can be tied to the
class via boost::bind(), using either copy semantic, pointers or
references (boost::ref()).

DBusServer extends upon the Timeout class by handling the ownership.
The Timeout class is deleted once the callback indicates that it no
longer wants to be called.

Most of the existing g_timeout_add_seconds() usage should be replaced
with this new class. The main point however is the BMC #7766, delayed
deletion of expired sessions.