synccompare: relax TZID simplification
For "compatible" timezone definitions there is no need to use exactly
the same TZID. Because comparing the definition is tricky, the script
fell back to normalizing TZIDs with a location at the end.

This location was not detected explicitly, but rather by the fact that
the TZID looked like an Olson one. TZID=Europe/Berlin thus did not
match.  Now the patch recognises locations (taken from
/usr/share/zoneinfo on Linux) and thus accepts such a TZID.