syncevo-dbus-server: incorrect Presence status after config change (BMC #8453)
This showed up in the sync-ui as "'Sync Now' button active after
creating a config while offline".

The root cause was a logical flaw in the syncevo-dbus-server: on
config change, it invalidated a cache of current
configs. Server::CheckPresence() then rebuilt and used that cache once
needed, but always so that all transports were enabled.

The solution is to rebuild the cache in PresenceStatus::init() so that
it matches the current network
state. PresenceStatus::updatePresenceStatus() does the same thing
before changing that state, which will include sending out the
necessary signals.