CalDAVSource: more efficient implementation of item listing
Implemented with a single REPORT request, instead of one query for
resource ID/etag pairs followed by one GET per item.

In theory also less data needs to be transmitted, because we only ask
for the minimum number of iCalendar 2.0 properties required at that
point. In practice, Google delivers full iCalendar 2.0 items. Because
we cannot rely on that, we do not populate the cache with these and
have to fetch those anew which are later involved in operations.

Overall this should still be more efficient.