MapSyncSource: provide Synthesis API pointer to sub source
Some operations in the sub sync source (like getDescription()) depend
on the Synthesis API pointer. This pointer is maintained in the
wrapping MapSyncSource. This patch makes that pointer available by
deriving SubSyncSource from SyncSourceBase and implementing the
getSynthesisAPI() inherited from it by calling the parent.

SyncSourceBase defines getSynthesisAPI(), so classes derived from
SyncSource and SubSyncSource inherit it. However, the implementation
is ambiguous, so such classes must define that they want the implementation
from SubSyncSource like this:
   SDKInterface *getSynthesisAPI() const { return SubSyncSource::getSynthesisAPI(); }