SyncConfig: make temporary changes to SyncSourceNodes properties survive as long as SyncConfig exists
Previously, SyncSourceNodes and the FilterConfigNodes with temporary
property values were created each time getSyncSourceNodes() was
called. This implied that such temporary changes did not survive across
multiple getSyncSourceNodes() calls.

This is an unexpected pitfall, which showed up again in
LocalTransportAgent ("disable all sources, then enable the ones that
are needed" failed because "sync = disabled" them was forgotten).

This patch solves this issue by keeping the full set of nodes around,
including the FilterConfigNodes with temporary changes, and returning
them in getSyncSourceNodes().

The interaction with setConfigFilter() is resolved such that calling
that function invalidates the cached nodes. They will be created
anew with the new filter applied, as before.