SyncSource: moved read-only access to operations into SyncSourceBase
This change is necessary for MapSyncSource, which needs to access
the operations via its SubSyncSource pointer, which is derived from
SyncSourceBase instead of SyncSource (to avoid double inheritance
of members).

Together with getOperations(), the definition also had to be moved,
which is such a big change that "diff" shows the change as "moving code
around Operations" instead of "moving Operations".