KCalExtended: avoid partial loading to work around issues with it
Partial loading can lead to a calendar which is in an inconsistent
state (child event exists in storage, but cannot be loaded to delete
it, BMC #6061).

This patch solves the issue with brute-force: load the complete calendar
once it is opened. This is less efficient, but better than incorrect
operation.