KCalExtended: workaround for importing custom time zones
This is an incomplete fix for importing custom time zones (BMC #8895).
It works as long as there is no TZID conflict. When a new time zone
definition uses the same TZID as another time zone, the old definition
is used, which may be incorrect. A full solution should use a new
mKCal API function (BMC #8604), which does not exist yet and still
needs to be discussed.