SYNCEVOLUTION_DEBUG: print DEBUG messages during sync if set
The code setting up a sync session disabled printing of DEBUG
messages to the console, even if it was enabled earlier because
of SYNCEVOLUTION_DEBUG.

Now check for SYNCEVOLUTION_DEBUG before making that change.
Not sure whether it is the right solution (perhaps just keep
the old level?), but it is known to have the desired effect.