local sync + BT: changed how credential checking is turned off
Resetting username/password temporarily was a hack which
had undesired side effects for local sync (credentials
were not available for other side of sync).

Changed this so that disabling of credential checking is done
at the point where it is needed: when creating Synthesis XML
config. Username/password settings are left untouched.