Buteo sync: implemented necessary functions to run sync
Interfacing with the Cmdline class turned out to be the easiest
way of implementing the functionality. It is a bit indirect, but
replicating the logic for setting up and updating the config
would have required more code.

Currently the config is updated each time the plugin runs. Should
change that so that an existing config is left unchanged.

Error handling is incomplete.