config handling: old-style configs not found anymore (BMC #9381)
Running a sync with SyncEvolution > 1.0.99.6 and a configuration created with
SyncEvolution <1.0 works once, then fails with "No configuration for ... found"
("[SyncEvolution] Problem after upgrade - old-style config not found in 1.1").

The root cause is this commit:

commit 1b622407085b39cfea0fd23686afe7d95f678fa7
Author: Patrick Ohly <patrick.ohly@intel.com>
Date:   Mon Aug 2 14:27:48 2010 +0200

    command line: allow creating contexts without peers

That patch incorrectly creates the "peers" directory also for configs
without a context. The bug was introduced because FileConfigTree only had
an ambiguously named boolean. Fixed by distinguishing behavior based on
the SyncContext enum.