StringDataBlob: implemented write() method
The method is needed to convert a SyncReport into an in-memory
string.

The problem is that it is uncertain when writing is complete.
This patch solves this by intercepting the deletion  of the ostream
via a boost::shared_ptr destructor which then copies the content
of the stream into the shared data buffer.