LocalTransportAgent: added transmission of child's SyncReport
The SyncReport is transmitted via an additional socket pair. This is
necessary because the one for messages is shut down to notify the peer
of a premature shutdown.

The parent uses the information obtained from the SyncReport to update
its own status: first the error is logged (thus also setting the
"first error seen" part of the parent), then an exception with a bad child
status is thrown (interrupting anything the parent was doing and, if it
was the first error, setting the overall status of the sync).